refactor(studio): 分离观测收件箱复核呈现
